falta revision final de todo

+++ Proyecto-2/ordenes-service/src/main/java/cl/duoc/ordenes_service/dto/ProductoDTO.java
@@ -7,6 +7,6 @@ public class ProductoDTO {
 
     private Long id;
     private String nombre;
-    private Double precio;
+    private Integer precio;
     private int stock;
 }
